+++ fabric/src/main/java/net/frozenblock/lib/data/frozenlib/FrozenLibRegistryProvider.java
@@ -33,6 +33,7 @@ final class FrozenLibRegistryProvider extends FabricDynamicRegistryProvider {
 	@Override
 	protected void configure(HolderLookup.Provider registries, Entries entries) {
 		entries.addAll(asLookup(entries.getLookup(FrozenLibRegistries.FIRE_TYPE)));
+		entries.addAll(asLookup(entries.getLookup(FrozenLibRegistries.VARIANT_SPAWN_INJECTION)));
 	}
 
 	public static <T> HolderLookup.RegistryLookup<T> asLookup(HolderGetter<T> getter) {